Backend : ajout page de garde

+++ database/migrations/2026_02_04_145738_create_dossiers_table.php
@@ -11,13 +11,15 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('dossiers', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('dao_id')->constrained('daos')->onDelete('cascade');
-            $table->string('nom_dossier');
-            $table->enum('statut', ['en_cours', 'termine', 'genere'])->default('en_cours');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('dossiers')) {
+            Schema::create('dossiers', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('dao_id')->constrained('daos')->onDelete('cascade');
+                $table->string('nom_dossier');
+                $table->enum('statut', ['en_cours', 'termine', 'genere'])->default('en_cours');
+                $table->timestamps();
+            });
+        }
     }
