+++ database/migrations/2026_02_04_150203_create_bordereaux_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('bordereaux', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('dossier_document_id')->constrained('dossier_documents')->onDelete('cascade');
-            $table->string('titre');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('bordereaux')) {
+            Schema::create('bordereaux', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('dossier_document_id')->constrained('dossier_documents')->onDelete('cascade');
+                $table->string('titre');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
